+++ src/UI/SignTransactionsModals/components/index.tsx
@@ -1,0 +1,4 @@
+export * from './ConfirmationScreen';
+export * from './DeviceConfirmationScreen';
+export * from './confirmationScreen.types';
+export * from './TransactionStatusToast';